+++ crates/storage/tests/chaos_questdb_lifecycle.rs
@@ -61,7 +61,7 @@ use tickvault_storage::tick_persistence::TickPersistenceWriter;
 /// Builds a minimal `ParsedTick` with the given security ID and price.
 fn make_tick(id: u32, price: f32) -> ParsedTick {
     ParsedTick {
-        security_id: id,
+        security_id: u64::from(id),
         exchange_segment_code: 1, // NSE_EQ
         last_traded_price: price,
         last_trade_quantity: 100,
